% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footer.R
\name{add_hf_content_type}
\alias{add_hf_content_type}
\title{Add Footer or Header Content Type}
\usage{
add_hf_content_type(doc, filename, hf_type)
}
\arguments{
\item{doc}{officer document}

\item{filename}{Filename (e.g., "footer1.xml", "header1.xml")}

\item{hf_type}{"footer" or "header"}
}
\description{
Add Footer or Header Content Type
}
\keyword{internal}
