% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footer.R
\name{add_hf_reference}
\alias{add_hf_reference}
\title{Add Footer or Header Reference to sectPr}
\usage{
add_hf_reference(doc, rid, type = "default", hf_type = "footer")
}
\arguments{
\item{doc}{officer document}

\item{rid}{The relationship ID}

\item{type}{"default", "first", or "even"}

\item{hf_type}{"footer" or "header"}
}
\description{
Add Footer or Header Reference to sectPr
}
\keyword{internal}
