% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footer.R
\name{add_hf_relationship}
\alias{add_hf_relationship}
\title{Add Footer or Header Relationship}
\usage{
add_hf_relationship(doc, target, hf_type)
}
\arguments{
\item{doc}{officer document}

\item{target}{Filename (e.g., "footer1.xml", "header1.xml")}

\item{hf_type}{"footer" or "header"}
}
\value{
The relationship ID (e.g., "rId31")
}
\description{
Adds a relationship to officer's internal relationship object.
}
\keyword{internal}
