% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footer.R
\name{add_title_page_setting}
\alias{add_title_page_setting}
\title{Add Title Page Setting}
\usage{
add_title_page_setting(doc)
}
\arguments{
\item{doc}{officer document}
}
\description{
Adds \verb{<w:titlePg/>} element to sectPr to enable different first page
headers/footers.
}
\keyword{internal}
