% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_styles.R
\name{apply_formal_list_style}
\alias{apply_formal_list_style}
\title{Apply Formal List Style (Internal)}
\usage{
apply_formal_list_style(docx_path)
}
\arguments{
\item{docx_path}{Path to the DOCX file}
}
\value{
The path (invisibly)
}
\description{
Modifies numbering.xml to use 1./a./i. format for numbered lists.
}
\keyword{internal}
