% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{build_comments_xml}
\alias{build_comments_xml}
\title{Build comments.xml content from comment data}
\usage{
build_comments_xml(comments)
}
\arguments{
\item{comments}{Named list of comment data}
}
\value{
List with \code{xml} (character string) and \code{para_ids} (named vector mapping comment ID to paraId)
}
\description{
Generates valid OpenXML for word/comments.xml from a list of comments.
Assigns paraIds to each comment for use with commentsExtended.xml.
}
\keyword{internal}
