% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{build_people_xml}
\alias{build_people_xml}
\title{Build people.xml content}
\usage{
build_people_xml(comments)
}
\arguments{
\item{comments}{Named list of comment data}
}
\value{
Character string containing people.xml content
}
\description{
Generates OpenXML for word/people.xml with unique authors from comments.
}
\keyword{internal}
