% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_inject.R
\name{build_table_cell}
\alias{build_table_cell}
\title{Build a Table Cell XML}
\usage{
build_table_cell(text, bold = FALSE)
}
\arguments{
\item{text}{Cell text content.}

\item{bold}{Whether to make the text bold.}
}
\value{
Character string containing the cell XML.
}
\description{
Build a Table Cell XML
}
\keyword{internal}
