% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_zotero.R
\name{build_zotero_pref_xml}
\alias{build_zotero_pref_xml}
\title{Build ZOTERO_PREF Field Code XML}
\usage{
build_zotero_pref_xml(
  style_id = "http://www.zotero.org/styles/vancouver",
  data_version = "4",
  field_type = "Fields",
  store_references = TRUE,
  journal_abbreviations = TRUE
)
}
\arguments{
\item{style_id}{CSL style ID (e.g., "http://www.zotero.org/styles/apa")}

\item{data_version}{Zotero data version (default: "4")}

\item{field_type}{"Fields" or "Bookmarks" (default: "Fields")}

\item{store_references}{Whether to store references in document (default: TRUE)}

\item{journal_abbreviations}{Whether Zotero should auto-abbreviate journal
names in the rendered bibliography (default: TRUE)}
}
\value{
Character string containing the field code XML
}
\description{
Generates the XML for a ZOTERO_PREF field code that can be injected
into a DOCX document. This is needed when the preferences field is
missing or corrupted.
}
\examples{
\dontrun{
# Generate APA style preferences
pref_xml <- build_zotero_pref_xml(
  style_id = "http://www.zotero.org/styles/apa"
)
}

}
\keyword{internal}
