% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_structure.R
\name{check_ids_in_file}
\alias{check_ids_in_file}
\title{Check IDs in a single XML file}
\usage{
check_ids_in_file(xml_path, id_requirements, ns, file_label)
}
\arguments{
\item{xml_path}{Path to XML file}

\item{id_requirements}{Named list of element -> attribute mappings}

\item{ns}{Namespace vector}

\item{file_label}{Label for error messages}
}
\value{
List with errors
}
\description{
Check IDs in a single XML file
}
\keyword{internal}
