% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revisions.R
\name{count_revisions}
\alias{count_revisions}
\title{Count revisions by type}
\usage{
count_revisions(revisions)
}
\arguments{
\item{revisions}{List from extract_revisions()}
}
\value{
Named vector with counts by type
}
\description{
Count revisions by type
}
