% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_injector.R
\name{css_to_inches}
\alias{css_to_inches}
\title{CSS to Inches}
\usage{
css_to_inches(val_str)
}
\arguments{
\item{val_str}{CSS string (e.g. "1in").}
}
\value{
Numeric inches.
}
\description{
CSS to Inches
}
\keyword{internal}
