% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css_injection.R
\name{css_to_pPr}
\alias{css_to_pPr}
\title{Convert CSS Properties to Word Paragraph Properties}
\usage{
css_to_pPr(css_props)
}
\arguments{
\item{css_props}{Named list of CSS properties}
}
\value{
List of Word paragraph properties
}
\description{
Convert CSS Properties to Word Paragraph Properties
}
\keyword{internal}
