% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docstyle-package.R
\docType{package}
\name{docstyle-package}
\alias{docstyle}
\alias{docstyle-package}
\title{docstyle: Style-as-Code for Scientific Publishing}
\description{
Generate Word, PDF, HTML, and PowerPoint assets from a single CSS + YAML configuration. Provides round-trip collaboration support for scientific publishing workflows, allowing reviewer edits from Word to be imported back into Quarto/Markdown source files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DougManuel/docstyle}
  \item Report bugs at \url{https://github.com/DougManuel/docstyle/issues}
}

}
\author{
\strong{Maintainer}: Douglas Manuel \email{dmanuel@ohri.ca}

}
\keyword{internal}
