% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_docstyle.R
\name{docstyle_version}
\alias{docstyle_version}
\title{Get docstyle Extension Version}
\usage{
docstyle_version(path = ".")
}
\arguments{
\item{path}{Path to the Quarto project directory. Defaults to current
working directory.}
}
\value{
A list with version information, or NULL if not installed.
}
\description{
Returns version information for the installed docstyle extension.
}
\examples{
\dontrun{
docstyle_version()
}

}
