% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{extract_comments}
\alias{extract_comments}
\title{Extract comments from a Word document}
\usage{
extract_comments(docx_path)
}
\arguments{
\item{docx_path}{Path to .docx file}
}
\value{
Named list of comment data, keyed by comment ID
}
\description{
Parses word/comments.xml from a DOCX file and extracts all comment metadata
including author, date, content, and threading information. Also reads
commentsExtended.xml for threading (parent_id) and resolved status (done).
}
\examples{
\dontrun{
comments <- extract_comments("document.docx")
comments[["1"]]$author
comments[["1"]]$content
comments[["1"]]$done
}

}
