% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revisions.R
\name{extract_initials}
\alias{extract_initials}
\title{Extract initials from author name}
\usage{
extract_initials(author)
}
\arguments{
\item{author}{Author name string}
}
\value{
Initials string
}
\description{
Extract initials from author name
}
\keyword{internal}
