% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revisions.R
\name{extract_revisions}
\alias{extract_revisions}
\title{Extract revisions (track changes) from a Word document}
\usage{
extract_revisions(docx_path)
}
\arguments{
\item{docx_path}{Path to .docx file}
}
\value{
Named list of revision data, keyed by revision ID
}
\description{
Parses document.xml from a DOCX file and extracts all tracked changes
including insertions (w:ins) and deletions (w:del) with their metadata.
}
\examples{
\dontrun{
revisions <- extract_revisions("document.docx")
revisions[["rev_1"]]$type
revisions[["rev_1"]]$author
}

}
