% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inject_zotero.R
\name{find_last_occurrence}
\alias{find_last_occurrence}
\title{Find Last Occurrence of a String}
\usage{
find_last_occurrence(haystack, needle)
}
\arguments{
\item{haystack}{String to search in}

\item{needle}{String to find}
}
\value{
Position of last occurrence, or -1 if not found
}
\description{
Find Last Occurrence of a String
}
\keyword{internal}
