% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_comments.R
\name{generate_html_report}
\alias{generate_html_report}
\title{Generate HTML report (internal)}
\usage{
generate_html_report(result, docx_path)
}
\description{
Generate HTML report (internal)
}
\keyword{internal}
