% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_comments.R
\name{generate_text_report}
\alias{generate_text_report}
\title{Generate text report (internal)}
\usage{
generate_text_report(result, docx_path)
}
\description{
Generate text report (internal)
}
\keyword{internal}
