% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_integration.R
\name{get_docstyle_config}
\alias{get_docstyle_config}
\title{Get Docstyle Configuration}
\usage{
get_docstyle_config(quarto_config)
}
\arguments{
\item{quarto_config}{List from \code{read_quarto_config}.}
}
\value{
List containing docstyle configuration or NULL.
}
\description{
Extracts the \code{docstyle} block from the Quarto configuration.
}
\keyword{internal}
