% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_layout.R
\name{get_page_dimensions}
\alias{get_page_dimensions}
\title{Get Page Dimensions in Twips}
\usage{
get_page_dimensions(size, orientation = "portrait")
}
\arguments{
\item{size}{Page size name (letter, a4, legal)}

\item{orientation}{portrait or landscape}
}
\value{
List with width and height in twips
}
\description{
Get Page Dimensions in Twips
}
\keyword{internal}
