% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_layout.R
\name{get_pandoc_reference_doc}
\alias{get_pandoc_reference_doc}
\title{Get Pandoc's Default Reference Document}
\usage{
get_pandoc_reference_doc()
}
\value{
Path to the extracted reference.docx
}
\description{
Extracts Pandoc's built-in reference.docx to a temp location.
}
\keyword{internal}
