% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pandoc_version}
\alias{get_pandoc_version}
\title{Get Pandoc version}
\usage{
get_pandoc_version()
}
\value{
A \code{numeric_version} object, or NULL if Pandoc is not found.
}
\description{
Get Pandoc version
}
\keyword{internal}
