% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_injector.R
\name{get_styles_xml}
\alias{get_styles_xml}
\title{Get Styles XML from Docx}
\usage{
get_styles_xml(doc)
}
\arguments{
\item{doc}{An \code{officer::read_docx} object.}
}
\value{
An \code{xml2} document object representing the styles.
}
\description{
Accesses the internal \code{word/styles.xml} of an \code{officer} document object.
}
\keyword{internal}
