% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_docstyle.R
\name{has_docstyle}
\alias{has_docstyle}
\title{Check if docstyle Extension is Installed}
\usage{
has_docstyle(path = ".")
}
\arguments{
\item{path}{Path to the Quarto project directory. Defaults to current
working directory.}
}
\value{
Logical. TRUE if extension is installed, FALSE otherwise.
}
\description{
Checks whether the docstyle Quarto extension is installed in a project.
}
\examples{
\dontrun{
if (!has_docstyle()) {
  use_docstyle()
}
}

}
