% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css_injection.R
\name{inject_pPr_to_style}
\alias{inject_pPr_to_style}
\title{Inject Paragraph Properties into Style Node}
\usage{
inject_pPr_to_style(style_node, pPr, ns)
}
\arguments{
\item{style_node}{XML style node}

\item{pPr}{List of paragraph properties}

\item{ns}{XML namespaces}
}
\description{
Inject Paragraph Properties into Style Node
}
\keyword{internal}
