% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css_injection.R
\name{inject_rPr_to_style}
\alias{inject_rPr_to_style}
\title{Inject Run Properties into Style Node}
\usage{
inject_rPr_to_style(style_node, rPr, ns)
}
\arguments{
\item{style_node}{XML style node}

\item{rPr}{List of run properties from css_to_rPr()}

\item{ns}{XML namespaces}
}
\description{
Handles rPr format from css_parser.R where properties are lists
(e.g., \code{rPr$b <- list()} for bold, \code{rPr$sz <- list(val = "24")}).
}
\keyword{internal}
