% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_injector.R
\name{inject_style_node}
\alias{inject_style_node}
\title{Inject Style into XML}
\usage{
inject_style_node(styles_xml, style_node)
}
\arguments{
\item{styles_xml}{The xml document (from \code{get_styles_xml}).}

\item{style_node}{The new style node (from \code{create_style_node}).}
}
\description{
Adds or updates a style node in the styles XML document.
}
\keyword{internal}
