% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchor_assembly.R
\name{is_text_box}
\alias{is_text_box}
\title{Check if a paragraph contains a text box}
\usage{
is_text_box(para, ns)
}
\arguments{
\item{para}{xml2 node for w:p}

\item{ns}{Named character vector of XML namespaces}
}
\value{
Logical
}
\description{
Detects \code{wp:anchor} containing \code{wps:txbx} but NOT \code{pic:pic}.
}
\keyword{internal}
