% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_zotero_running}
\alias{is_zotero_running}
\title{Check if Zotero with Better BibTeX is Running}
\usage{
is_zotero_running()
}
\value{
TRUE if Zotero with Better BibTeX is responding, FALSE otherwise.
}
\description{
Attempts to connect to Better BibTeX's local HTTP endpoint.
}
\details{
Better BibTeX exposes an HTTP API at \verb{http://127.0.0.1:23119/better-bibtex/}.
This function attempts a simple request to check connectivity.
}
