% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css_injection.R
\name{map_selector_to_word_style}
\alias{map_selector_to_word_style}
\title{Map CSS Selector to Word Style}
\usage{
map_selector_to_word_style(selector)
}
\arguments{
\item{selector}{CSS selector}
}
\value{
List with id and name, or NULL if unmappable
}
\description{
Map CSS Selector to Word Style
}
\keyword{internal}
