% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_matcher.R
\name{normalize_items}
\alias{normalize_items}
\title{Normalize Citation Items}
\usage{
normalize_items(items, source)
}
\arguments{
\item{items}{List of CSL items.}

\item{source}{Label ("word" or "bib").}
}
\value{
Data frame.
}
\description{
Converts a list of CSL items to a data frame for comparison.
}
\keyword{internal}
