% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css_reader.R
\name{parse_css_content}
\alias{parse_css_content}
\title{Parse CSS Content}
\usage{
parse_css_content(text)
}
\arguments{
\item{text}{Raw CSS string.}
}
\value{
Named list of styles with a \code{page} attribute containing @page rules.
}
\description{
Parse CSS Content
}
\keyword{internal}
