% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css_reader.R
\name{parse_css_properties}
\alias{parse_css_properties}
\title{Parse CSS Properties Block}
\usage{
parse_css_properties(text)
}
\arguments{
\item{text}{Content inside curly braces from a CSS rule.}
}
\value{
Named list of properties.
}
\description{
Parse CSS Properties Block
}
\keyword{internal}
