% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_layout.R
\name{parse_docstyle_config}
\alias{parse_docstyle_config}
\title{Parse docstyle Configuration}
\usage{
parse_docstyle_config(config_path)
}
\arguments{
\item{config_path}{Path to YAML file or a list}
}
\value{
Parsed configuration list
}
\description{
Parse docstyle Configuration
}
\keyword{internal}
