% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css_injection.R
\name{parse_font_shorthand}
\alias{parse_font_shorthand}
\title{Parse CSS Font Shorthand}
\usage{
parse_font_shorthand(font_str)
}
\arguments{
\item{font_str}{The font shorthand string}
}
\value{
List with weight, size, family components
}
\description{
Parse CSS Font Shorthand
}
\keyword{internal}
