% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css_reader.R
\name{parse_page_properties}
\alias{parse_page_properties}
\title{Parse CSS @page Properties}
\usage{
parse_page_properties(props)
}
\arguments{
\item{props}{Named list of CSS properties from @page rule.}
}
\value{
List with docstyle page configuration.
}
\description{
Converts CSS @page properties to docstyle page configuration format.
}
\keyword{internal}
