% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib_reader.R
\name{read_bib_as_csl}
\alias{read_bib_as_csl}
\title{Read BibTeX as CSL-JSON}
\usage{
read_bib_as_csl(bib_path)
}
\arguments{
\item{bib_path}{Path to the .bib file.}
}
\value{
A list of CSL-JSON items.
}
\description{
Converts a BibTeX file to CSL-JSON using Pandoc.
This ensures robust parsing and normalization of citation data.
}
\keyword{internal}
