% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{read_comments_json}
\alias{read_comments_json}
\title{Read comments from JSON sidecar file}
\usage{
read_comments_json(path)
}
\arguments{
\item{path}{Path to comments.json}
}
\value{
Named list of comment data
}
\description{
Read comments from JSON sidecar file
}
\keyword{internal}
