% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revisions.R
\name{read_revisions_json}
\alias{read_revisions_json}
\title{Read revisions from JSON sidecar file}
\usage{
read_revisions_json(path)
}
\arguments{
\item{path}{Path to revisions.json}
}
\value{
Named list of revision data
}
\description{
Read revisions from JSON sidecar file
}
\keyword{internal}
