% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{repack_docx}
\alias{repack_docx}
\title{Repack a DOCX from extracted directory}
\usage{
repack_docx(temp_dir, docx_path)
}
\arguments{
\item{temp_dir}{Path to extracted DOCX directory}

\item{docx_path}{Output path for DOCX file}
}
\description{
Repack a DOCX from extracted directory
}
\keyword{internal}
