% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{scan_used_comment_ids}
\alias{scan_used_comment_ids}
\title{Scan document.xml for used comment IDs}
\usage{
scan_used_comment_ids(docx_path)
}
\arguments{
\item{docx_path}{Path to .docx file}
}
\value{
Character vector of comment IDs found in document
}
\description{
Searches the rendered document.xml for w:commentRangeStart elements
to determine which comment IDs are actually referenced.
}
\keyword{internal}
