% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revisions.R
\name{summarise_revisions_by_author}
\alias{summarise_revisions_by_author}
\title{Summarise revisions by author}
\usage{
summarise_revisions_by_author(revisions)
}
\arguments{
\item{revisions}{List from extract_revisions()}
}
\value{
Data frame with author-level summary
}
\description{
Summarise revisions by author
}
