% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{update_content_types_for_comments}
\alias{update_content_types_for_comments}
\title{Update Content_Types.xml to include comments files}
\usage{
update_content_types_for_comments(
  temp_dir,
  has_extended = FALSE,
  has_people = FALSE
)
}
\arguments{
\item{temp_dir}{Path to extracted DOCX directory}

\item{has_extended}{Whether to add commentsExtended.xml override}

\item{has_people}{Whether to add people.xml override}
}
\description{
Adds overrides for comments.xml and optionally commentsExtended.xml and people.xml.
}
\keyword{internal}
