% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{update_rels_for_comments}
\alias{update_rels_for_comments}
\title{Update document.xml.rels to include comments relationships}
\usage{
update_rels_for_comments(temp_dir, has_extended = FALSE, has_people = FALSE)
}
\arguments{
\item{temp_dir}{Path to extracted DOCX directory}

\item{has_extended}{Whether to add commentsExtended.xml relationship}

\item{has_people}{Whether to add people.xml relationship}
}
\description{
Update document.xml.rels to include comments relationships
}
\keyword{internal}
