% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_injector.R
\name{write_styles_xml}
\alias{write_styles_xml}
\title{Write Styles XML to Docx}
\usage{
write_styles_xml(doc, xml)
}
\arguments{
\item{doc}{An \code{officer::read_docx} object.}

\item{xml}{An \code{xml2} document object to write.}
}
\description{
Writes the modified styles XML back to the \code{officer} document object's temp directory.
}
\keyword{internal}
