% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xml_escape}
\alias{xml_escape}
\title{XML Escape}
\usage{
xml_escape(text)
}
\arguments{
\item{text}{Text to escape}
}
\value{
Escaped text
}
\description{
Escapes special characters for XML text content.
}
\keyword{internal}
