% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xml_escape_attr}
\alias{xml_escape_attr}
\title{XML Escape for Attribute Values}
\usage{
xml_escape_attr(text)
}
\arguments{
\item{text}{Text to escape for use in XML attributes}
}
\value{
Escaped text safe for XML attributes
}
\description{
Escapes all five XML entities for use in attribute values.
}
\keyword{internal}
